First cut of CRC/LFSR move to main lib from Crypto
This is more a generic component set, so we need to move it into lib
instead of Crypto so more components can leverage LFSR, etc. CRC
support.  Compiles.  SEEMS to act right on first blush.  Need to
fix VexRiscv and Crypto for anyting this breaks next.  Committing
to ensure we preserve the first cut here on this.